+++ src/modules/simulation/gz_bridge/GZMixingInterfaceServo.hpp
@@ -49,8 +49,7 @@ public:
 		_node(node)
 	{}
 
-	bool updateOutputs(uint16_t outputs[MAX_ACTUATORS],
-			   unsigned num_outputs, unsigned num_control_groups_updated) override;
+	bool updateOutputs(float outputs[MAX_ACTUATORS], unsigned num_outputs, unsigned num_control_groups_updated) override;
 
 	MixingOutput &mixingOutput() { return _mixing_output; }
